# Issue 315 elementwise function jac