--- a/pybamm/__init__.py
+++ b/pybamm/__init__.py
@@ -80,7 +80,6 @@ from .expression_tree.unary_operators import (
     BoundaryValue,
     Integral,
     IndefiniteIntegral,
-    Diagonal,
     grad,
     div,
     surf,