# Description

- Compute Function jacobian elementwise 
- Remove `Diagonal` nodes and use elementwise multiplication/division instead 
- Remove `SparseStack.sparsify()` (no longer needed)

The `evaluate` hack from `Multiplication` is now used in `Division` as well, ideally will be removed by #320 
Part of #315 

## Speed-ups (tests on DFN)

Before:
```
In [1]: %timeit jac.evaluate(0,y0)                                                                                                            
61 ms ± 447 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [2]: %timeit jac.evaluate(0,y0,{})                                                                                                         
29 ms ± 219 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %timeit self.integrate( residuals, y0, t_eval, events=events, mass_matrix=model.mass_matrix.entries, jacobian=jacobian)               
2.72 s ± 7.96 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %prun jac.evaluate(0,y0)
90456 function calls (89453 primitive calls) in 0.105 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      301    0.005    0.000    0.014    0.000 compressed.py:133(check_format)
      985    0.005    0.000    0.011    0.000 sputils.py:120(get_index_dtype)
     5424    0.005    0.000    0.005    0.000 {built-in method numpy.array}
 1073/828    0.004    0.000    0.012    0.000 tracer.py:35(f_wrapped)
     1970    0.003    0.000    0.003    0.000 getlimits.py:497(__init__)
       67    0.003    0.000    0.012    0.000 dia.py:348(tocoo)
    639/6    0.003    0.000    0.105    0.017 binary_operators.py:560(evaluate)
       82    0.002    0.000    0.006    0.000 coo.py:261(_check)
  301/261    0.002    0.000    0.025    0.000 compressed.py:25(__init__)
       45    0.002    0.000    0.015    0.000 core.py:17(backward_pass)
     1825    0.002    0.000    0.005    0.000 tracer.py:58(f_wrapped)
      324    0.002    0.000    0.002    0.000 {method 'reduce' of 'numpy.ufunc' objects}
      319    0.002    0.000    0.004    0.000 compressed.py:1077(prune)
      458    0.002    0.000    0.003    0.000 sputils.py:266(check_shape)
      245    0.002    0.000    0.004    0.000 core.py:28(__init__)
     1073    0.001    0.000    0.002    0.000 tracer.py:65(find_top_boxed_args)
      891    0.001    0.000    0.002    0.000 tracer.py:115(toposort)
       63    0.001    0.000    0.007    0.000 construct.py:65(diags)
      774    0.001    0.000    0.004    0.000 numpy_vjps.py:548(unbroadcast)
     4170    0.001    0.000    0.002    0.000 {built-in method builtins.isinstance}
       65    0.001    0.000    0.016    0.000 compressed.py:486(_mul_sparse_matrix)
       80    0.001    0.000    0.011    0.000 coo.py:368(tocsr)
     3765    0.001    0.000    0.004    0.000 numeric.py:469(asarray)
      484    0.001    0.000    0.001    0.000 {built-in method numpy.empty}
      334    0.001    0.000    0.001    0.000 sputils.py:209(isshape)
       16    0.001    0.000    0.006    0.000 graphite_mcmb2528_ocp_Dualfoil.py:4(graphite_mcmb2528_ocp_Dualfoil)
     7342    0.001    0.000    0.001    0.000 {built-in method builtins.len}
      175    0.001    0.000    0.002    0.000 binary_operators.py:759(_binary_evaluate)
       82    0.001    0.000    0.009    0.000 coo.py:126(__init__)
     1312    0.001    0.000    0.001    0.000 compressed.py:105(getnnz)
     3747    0.001    0.000    0.001    0.000 tracer.py:144(<lambda>)
     1749    0.001    0.000    0.003    0.000 base.py:243(nnz)
```
After:
```
In [2]: %timeit jac.evaluate(0,y0)                                                                                                            
37.3 ms ± 955 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %timeit jac.evaluate(0,y0,{})                                                                                                         
17.7 ms ± 161 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [4]: %timeit self.integrate( residuals, y0, t_eval, events=events, mass_matrix=model.mass_matrix.entries, jacobian=jacobian)               
2.13 s ± 11.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: %prun jac.evaluate(0,y0)  
         53361 function calls (52413 primitive calls) in 0.064 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      160    0.003    0.000    0.008    0.000 compressed.py:133(check_format)
      542    0.003    0.000    0.006    0.000 sputils.py:120(get_index_dtype)
  796/551    0.003    0.000    0.008    0.000 tracer.py:35(f_wrapped)
    613/6    0.002    0.000    0.062    0.010 binary_operators.py:558(evaluate)
     2604    0.002    0.000    0.002    0.000 {built-in method numpy.array}
     1084    0.002    0.000    0.002    0.000 getlimits.py:497(__init__)
       66    0.002    0.000    0.005    0.000 coo.py:261(_check)
       74    0.002    0.000    0.014    0.000 compressed.py:355(multiply)
      248    0.001    0.000    0.001    0.000 {method 'reduce' of 'numpy.ufunc' objects}
  160/136    0.001    0.000    0.015    0.000 compressed.py:25(__init__)
      198    0.001    0.000    0.002    0.000 compressed.py:1077(prune)
      235    0.001    0.000    0.018    0.000 binary_operators.py:751(_binary_evaluate)
       22    0.001    0.000    0.005    0.000 bsr.py:472(tocoo)
      796    0.001    0.000    0.001    0.000 tracer.py:65(find_top_boxed_args)
       58    0.001    0.000    0.010    0.000 coo.py:368(tocsr)
      238    0.001    0.000    0.001    0.000 sputils.py:266(check_shape)
     2925    0.001    0.000    0.002    0.000 {built-in method builtins.isinstance}
       16    0.001    0.000    0.005    0.000 graphite_mcmb2528_ocp_Dualfoil.py:4(graphite_mcmb2528_ocp_Dualfoil)
      696    0.001    0.000    0.002    0.000 tracer.py:58(f_wrapped)
      611    0.001    0.000    0.001    0.000 {method 'reshape' of 'numpy.ndarray' objects}
       13    0.001    0.000    0.005    0.000 core.py:17(backward_pass)
       16    0.001    0.000    0.003    0.000 lico2_ocp_Dualfoil.py:4(lico2_ocp_Dualfoil)
      860    0.001    0.000    0.001    0.000 compressed.py:105(getnnz)
        9    0.001    0.000    0.002    0.000 construct.py:401(_compressed_sparse_stack)
      193    0.001    0.000    0.001    0.000 sputils.py:209(isshape)
     1837    0.001    0.000    0.002    0.000 numeric.py:469(asarray)
      257    0.001    0.000    0.001    0.000 {built-in method numpy.empty}
    66/62    0.001    0.000    0.007    0.000 coo.py:126(__init__)
     1973    0.001    0.000    0.001    0.000 tracer.py:144(<lambda>)
       34    0.001    0.000    0.011    0.000 compressed.py:1147(_binopt)
     4257    0.001    0.000    0.001    0.000 {built-in method builtins.len}
     1112    0.001    0.000    0.002    0.000 base.py:243(nnz)
```
Not sure why the first three lines are still being called, will investigate

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# Key checklist:

- [x] No style issues: `$ flake8`
- [x] All tests pass: `$ python run-tests.py --unit`
- [x] The documentation builds: `$ cd docs` and then `$ make clean; make html`

You can run all three at once, using `$ python run-tests.py --quick`.

## Further checks: 

- [x] Code is commented, particularly in hard-to-understand areas
- [x] Tests added that prove fix is effective or that feature works
- [ ] Any dependent changes have been merged and published in downstream modules